+++ app/scenes/Player/components/Upload/index.js
@@ -1,0 +1,3 @@
+import Upload from './Upload'
+
+export default Upload
